Submit2

+++ src - Original/main.cpp
@@ -60,7 +60,7 @@ int main()
           
           MeasurementPackage meas_package;
           istringstream iss(sensor_measurment);
-    	  double timestamp;
+    	  long long timestamp;
 
     	  // reads first element from the current line
     	  string sensor_type;
